requirements.txt
requests
python-dotenv
xmltodict
pytest
mock
colorama